+++ src/eigen/lobpcg_standard.cpp
@@ -1,5 +1,6 @@
 #include "../../include/monolish_blas.hpp"
 #include "../../include/monolish_eigen.hpp"
+#include "../../include/monolish_vml.hpp"
 #include "../internal/lapack/monolish_lapack.hpp"
 #include "../internal/monolish_internal.hpp"
 
@@ -7,74 +8,139 @@ namespace monolish {
 
 template <typename MATRIX, typename T>
 int standard_eigen::LOBPCG<MATRIX, T>::monolish_LOBPCG(
-    MATRIX &A, T &l, monolish::vector<T> &xinout) {
-  T norm;
+    MATRIX &A, vector<T> &l, matrix::Dense<T> &xinout) {
   Logger &logger = Logger::get_instance();
   logger.solver_in(monolish_func);
 
-  this->precond.create_precond(A);
-  // Algorithm following DOI:10.1007/978-3-319-69953-0_14
-  xinout[0] = 1.0;
-  xinout[1] = -1.0;
-  blas::nrm2(xinout, norm);
-  blas::scal(1.0 / norm, xinout);
-  monolish::matrix::Dense<T> wxp(3, A.get_col());
-  monolish::matrix::Dense<T> twxp(A.get_col(), 3);
-  monolish::matrix::Dense<T> WXP(3, A.get_row());
-  monolish::view1D<monolish::matrix::Dense<T>, T> w(wxp, 0, 1 * A.get_row());
-  monolish::view1D<monolish::matrix::Dense<T>, T> x(wxp, 1 * A.get_row(),
-                                                    2 * A.get_row());
-  monolish::view1D<monolish::matrix::Dense<T>, T> p(wxp, 2 * A.get_row(),
-                                                    3 * A.get_row());
-  monolish::view1D<monolish::matrix::Dense<T>, T> W(WXP, 0, 1 * A.get_row());
-  monolish::view1D<monolish::matrix::Dense<T>, T> X(WXP, 1 * A.get_row(),
-                                                    2 * A.get_row());
-  monolish::view1D<monolish::matrix::Dense<T>, T> P(WXP, 2 * A.get_row(),
-                                                    3 * A.get_row());
-  monolish::vector<T> vtmp1(A.get_row());
-  monolish::vector<T> vtmp2(A.get_row());
-
-  if (A.get_device_mem_stat() == true) {
-    monolish::util::send(wxp, twxp, WXP, vtmp1, vtmp2, xinout);
+  // size consistency check
+  // the size of xinout is (m, A.get_col())
+  if (A.get_col() != xinout.get_col()) {
+    throw std::runtime_error("error, A.col != x.col");
+  }
+  if (l.size() != xinout.get_row()) {
+    throw std::runtime_error("error, l.size != x.row");
   }
 
-  blas::copy(xinout, x);
-  // X = A x
-  blas::matvec(A, x, X);
-  // mu = (x, X)
-  T mu;
-  blas::dot(x, X, mu);
-  // w = X - mu x
-  blas::axpyz(-mu, x, X, w);
-  blas::nrm2(w, norm);
-  blas::scal(1.0 / norm, w);
+  this->precond.create_precond(A);
+
+  // Algorithm following DOI:10.1007/978-3-319-69953-0_14
+  // m is the number of eigenpairs to compute
+  const std::size_t m = l.size();
+  // n is the dimension of the original space
+  const std::size_t n = A.get_col();
+
+  // Internal memory
+  // currently 6m+(6m+3m+4) vectors are used
+  matrix::Dense<T> wxp(3 * m, n);
+  matrix::Dense<T> WXP(3 * m, n);
+  // TODO: wxp_p and WXP_p are not needed when m=1
+  matrix::Dense<T> wxp_p(3 * m, n);
+  matrix::Dense<T> WXP_p(3 * m, n);
+  // TODO: twxp is not needed when transpose matmul is supported
+  matrix::Dense<T> twxp(n, 3 * m);
+  // TODO: vtmp1 and vtmp2 are not needed when in-place preconditioner is used
+  //       and view1D is supported by preconditioners
+  monolish::vector<T> vtmp1(n);
+  monolish::vector<T> vtmp2(n);
+  // TODO: zero is not needed when view1D supports fill(T val)
+  monolish::vector<T> zero(n, 0.0);
+  // TODO: r is not needed when util::random_vector supports view1D
+  monolish::vector<T> r(n);
+
+  // view1Ds for calculation
+  std::vector<view1D<matrix::Dense<T>, T>> w;
+  std::vector<view1D<matrix::Dense<T>, T>> x;
+  std::vector<view1D<matrix::Dense<T>, T>> p;
+  std::vector<view1D<matrix::Dense<T>, T>> W;
+  std::vector<view1D<matrix::Dense<T>, T>> X;
+  std::vector<view1D<matrix::Dense<T>, T>> P;
+  std::vector<view1D<matrix::Dense<T>, T>> wp;
+  std::vector<view1D<matrix::Dense<T>, T>> xp;
+  std::vector<view1D<matrix::Dense<T>, T>> pp;
+  std::vector<view1D<matrix::Dense<T>, T>> Wp;
+  std::vector<view1D<matrix::Dense<T>, T>> Xp;
+  std::vector<view1D<matrix::Dense<T>, T>> Pp;
+  for (std::size_t i = 0; i < m; ++i) {
+    w.push_back(view1D<matrix::Dense<T>, T>(wxp, i * n, (i + 1) * n));
+    x.push_back(view1D<matrix::Dense<T>, T>(wxp, (m + i) * n, (m + i + 1) * n));
+    p.push_back(
+        view1D<matrix::Dense<T>, T>(wxp, (2 * m + i) * n, (2 * m + i + 1) * n));
+    W.push_back(view1D<matrix::Dense<T>, T>(WXP, i * n, (i + 1) * n));
+    X.push_back(view1D<matrix::Dense<T>, T>(WXP, (m + i) * n, (m + i + 1) * n));
+    P.push_back(
+        view1D<matrix::Dense<T>, T>(WXP, (2 * m + i) * n, (2 * m + i + 1) * n));
+    wp.push_back(view1D<matrix::Dense<T>, T>(wxp_p, i * n, (i + 1) * n));
+    xp.push_back(
+        view1D<matrix::Dense<T>, T>(wxp_p, (m + i) * n, (m + i + 1) * n));
+    pp.push_back(view1D<matrix::Dense<T>, T>(wxp_p, (2 * m + i) * n,
+                                             (2 * m + i + 1) * n));
+    Wp.push_back(view1D<matrix::Dense<T>, T>(WXP_p, i * n, (i + 1) * n));
+    Xp.push_back(
+        view1D<matrix::Dense<T>, T>(WXP_p, (m + i) * n, (m + i + 1) * n));
+    Pp.push_back(view1D<matrix::Dense<T>, T>(WXP_p, (2 * m + i) * n,
+                                             (2 * m + i + 1) * n));
+  }
+
+  // Preparing initial input to be orthonormal to each other
+  for (std::size_t i = 0; i < m; ++i) {
+    T minval = 0.0;
+    T maxval = 1.0;
+    util::random_vector(r, minval, maxval);
+    blas::copy(r, x[i]);
+    T norm;
+    blas::nrm2(x[i], norm);
+    blas::scal(1.0 / norm, x[i]);
+  }
+
+  if (A.get_device_mem_stat() == true) {
+    monolish::util::send(wxp, WXP, wxp_p, WXP_p, twxp, vtmp1, vtmp2, zero,
+                         xinout);
+  }
+
+  for (std::size_t i = 0; i < m; ++i) {
+    // X = A x
+    blas::matvec(A, x[i], X[i]);
+    // mu = (x, X)
+    T mu = blas::dot(x[i], X[i]);
+    // w = X - mu x, normalize
+    blas::axpyz(-mu, x[i], X[i], w[i]);
+    T norm = blas::nrm2(w[i]);
+    blas::scal(1.0 / norm, w[i]);
+    // W = A w
+    blas::matvec(A, w[i], W[i]);
+  }
 
   // B singular flag
   bool is_singular = false;
-  // W = A w
-  blas::matvec(A, w, W);
-  matrix::Dense<T> Sam(3, 3);
-  matrix::Dense<T> Sbm(3, 3);
-  vector<T> lambda(Sam.get_col());
+
+  matrix::Dense<T> Sam(3 * m, 3 * m);
+  matrix::Dense<T> Sbm(3 * m, 3 * m);
+  vector<T> lambda(3 * m);
+  if (A.get_device_mem_stat() == true) {
+    monolish::util::send(Sbm);
+  }
 
   for (std::size_t iter = 0; iter < this->get_maxiter(); iter++) {
     if (A.get_device_mem_stat() == true) {
-      monolish::util::send(Sam, Sbm, lambda);
+      monolish::util::send(Sam, lambda);
     }
     if (iter == 0 || is_singular) {
       // It is intended not to resize actual memory layout
       // and just use the beginning part of
       // (i.e. not touching {Sam,Sbm,wxp,twxp,WXP}.{val,nnz})
-      Sam.set_col(2);
-      Sam.set_row(2);
-      Sbm.set_col(2);
-      Sbm.set_row(2);
-      wxp.set_row(2);
-      twxp.set_col(2);
-      WXP.set_row(2);
+      Sam.set_col(2 * m);
+      Sam.set_row(2 * m);
+      Sbm.set_col(2 * m);
+      Sbm.set_row(2 * m);
+      wxp.set_row(2 * m);
+      wxp_p.set_row(2 * m);
+      twxp.set_col(2 * m);
+      WXP.set_row(2 * m);
+      WXP_p.set_row(2 * m);
     }
     if (A.get_device_mem_stat() == true) {
       wxp.nonfree_recv();
+      twxp.device_free();
     }
     twxp.transpose(wxp);
     if (A.get_device_mem_stat() == true) {
@@ -86,6 +152,7 @@ int standard_eigen::LOBPCG<MATRIX, T>::monolish_LOBPCG(
     // Sb = { w, x, p }^T { w, x, p }
     blas::matmul(wxp, twxp, Sbm);
 
+    // workaround on GPU; synchronize Sam, Sbm
     if (A.get_device_mem_stat() == true) {
       Sam.nonfree_recv();
       Sbm.nonfree_recv();
@@ -95,9 +162,16 @@ int standard_eigen::LOBPCG<MATRIX, T>::monolish_LOBPCG(
     const char jobz = 'V';
     const char uplo = 'L';
     int info = internal::lapack::sygvd(Sam, Sbm, lambda, 1, &jobz, &uplo);
-    // info==6 means order 3 of B is not positive definite, similar to step 0.
-    // therefore we set is_singular flag to true and restart the iteration step.
-    if (info == 6) {
+    // 5m < info <= 6m means order 3 of B is not positive definite, similar to
+    // step 0. therefore we set is_singular flag to true and restart the
+    // iteration step.
+    int lbound = 5 * m;
+    int ubound = 6 * m;
+    if (iter == 0 || is_singular) {
+      lbound = 3 * m;
+      ubound = 4 * m;
+    }
+    if (lbound < info && info <= ubound) {
       if (this->get_print_rhistory()) {
         *this->rhistory_stream << iter + 1 << "\t"
                                << "singular; restart the step" << std::endl;
@@ -105,100 +179,134 @@ int standard_eigen::LOBPCG<MATRIX, T>::monolish_LOBPCG(
       is_singular = true;
       iter--;
       continue;
-    } else if (info != 0) {
-      throw std::runtime_error("internal LAPACK sygvd returned error");
+    } else if (info != 0 && info < static_cast<int>(m)) {
+      std::string s = "sygvd returns ";
+      s += std::to_string(info);
+      s += ": internal LAPACK sygvd returned error";
+      throw std::runtime_error(s);
     }
     if (A.get_device_mem_stat() == true) {
       monolish::util::recv(Sam, lambda);
     }
-    std::size_t index = 0;
-    l = lambda[index];
 
-    // extract b which satisfies Aprime b = lambda_min b
-    size_t size_b = std::sqrt(Sam.get_nnz());
-    monolish::vector<T> b(size_b);
-    Sam.row(index, b);
+    // prepare previous step results
+    blas::copy(wxp, wxp_p);
+    blas::copy(WXP, WXP_p);
+    vector<T> residual(m);
+    for (std::size_t i = 0; i < m; ++i) {
+      // copy current eigenvalue results
+      l[i] = lambda[i];
 
-    if (iter == 0 || is_singular) {
-      // b[2] is not calculated so explicitly set to 0
-      b[2] = 0.0;
-    }
+      // extract eigenvector of Sa v = lambda Sb v
+      vector<T> b(3 * m);
+      if (iter == 0 || is_singular) {
+        b.resize(2 * m);
+      }
+      Sam.row(i, b);
 
-    // x = b[0] w + b[1] x + b[2] p, normalize
-    // p = b[0] w          + b[2] p, normalize
-    blas::scal(b[0], w);
-    blas::xpay(b[2], w, p);
-    blas::xpay(b[1], p, x);
+      if (iter == 0 || is_singular) {
+        b.resize(3 * m);
+        // b[2m]...b[3m-1] is not calculated so explicitly set to 0
+        for (std::size_t j = 2 * m; j < 3 * m; ++j) {
+          b[j] = 0.0;
+        }
+      }
+      blas::copy(zero, p[i]);
+      blas::copy(zero, x[i]);
+      blas::copy(zero, P[i]);
+      blas::copy(zero, X[i]);
+      for (std::size_t j = 0; j < m; ++j) {
+        // x[i] = \Sum_j b[j] w[j] + b[m+j] x[j] + b[2m+j] p[j], normalize
+        // p[i] = \Sum_j b[j] w[j]               + b[2m+j] p[j], normalize
+        blas::axpy(b[j], wp[j], p[i]);
+        blas::axpy(b[2 * m + j], pp[j], p[i]);
+        blas::axpy(b[j], wp[j], x[i]);
+        blas::axpy(b[2 * m + j], pp[j], x[i]);
+        blas::axpy(b[m + j], xp[j], x[i]);
 
-    // X = b[0] W + b[1] X + b[2] P, normalize with x
-    // P = b[0] W          + b[2] P, normalize with p
-    blas::scal(b[0], W);
-    blas::xpay(b[2], W, P);
-    blas::xpay(b[1], P, X);
-    T normp;
-    blas::nrm2(p, normp);
-    blas::scal(1.0 / normp, p);
-    blas::scal(1.0 / normp, P);
-    T normx;
-    blas::nrm2(x, normx);
-    blas::scal(1.0 / normx, x);
-    blas::scal(1.0 / normx, X);
+        // X[i] = \Sum_j b[j]W[j] + b[m+j]X[j] + b[2m+j]P[j], normalize with x
+        // P[i] = \Sum_j b[j]W[j]              + b[2m+j]P[j], normalize with p
+        blas::axpy(b[j], Wp[j], P[i]);
+        blas::axpy(b[2 * m + j], Pp[j], P[i]);
+        blas::axpy(b[j], Wp[j], X[i]);
+        blas::axpy(b[2 * m + j], Pp[j], X[i]);
+        blas::axpy(b[m + j], Xp[j], X[i]);
+      }
 
-    // w = X - lambda x
-    blas::axpyz(-l, x, X, w);
-    // apply preconditioner
-    blas::copy(w, vtmp2);
-    this->precond.apply_precond(vtmp2, vtmp1);
-    blas::copy(vtmp1, w);
+      T normp;
+      blas::nrm2(p[i], normp);
+      blas::scal(1.0 / normp, p[i]);
+      blas::scal(1.0 / normp, P[i]);
+      T normx;
+      blas::nrm2(x[i], normx);
+      blas::scal(1.0 / normx, x[i]);
+      blas::scal(1.0 / normx, X[i]);
 
-    // residual calculation
-    T residual;
-    blas::nrm2(w, residual);
-    if (this->get_print_rhistory()) {
-      *this->rhistory_stream << iter + 1 << "\t" << std::scientific << residual
-                             << std::endl;
+      // w[i] = X[i] - lambda x[i]
+      blas::axpyz(-l[i], x[i], X[i], w[i]);
+      // apply preconditioner
+      blas::copy(w[i], vtmp2);
+      this->precond.apply_precond(vtmp2, vtmp1);
+      blas::copy(vtmp1, w[i]);
+
+      // residual calculation
+      blas::nrm2(w[i], residual[i]);
+      if (this->get_print_rhistory()) {
+        *this->rhistory_stream << iter + 1 << "\t" << i << "\t"
+                               << std::scientific << residual[i] << std::endl;
+      }
+      if (std::isnan(residual[i])) {
+        for (std::size_t j = 0; j < m; ++j) {
+          view1D<matrix::Dense<T>, T> xinout_j(xinout, j * n, (j + 1) * n);
+          blas::copy(x[j], xinout_j);
+        }
+        logger.solver_out();
+        return MONOLISH_SOLVER_RESIDUAL_NAN;
+      }
+      // normalize w
+      blas::scal(1.0 / residual[i], w[i]);
+      // W = A w
+      blas::matvec(A, w[i], W[i]);
     }
 
     // early return when residual is small enough
-    if (residual < this->get_tol() && this->get_miniter() < iter + 1) {
-      blas::copy(x, xinout);
+    if (vml::max(residual) < this->get_tol() &&
+        this->get_miniter() < iter + 1) {
+      for (std::size_t i = 0; i < m; ++i) {
+        view1D<matrix::Dense<T>, T> xinout_i(xinout, i * n, (i + 1) * n);
+        blas::copy(x[i], xinout_i);
+      }
       logger.solver_out();
       return MONOLISH_SOLVER_SUCCESS;
     }
 
-    if (std::isnan(residual)) {
-      blas::copy(x, xinout);
-      logger.solver_out();
-      return MONOLISH_SOLVER_RESIDUAL_NAN;
-    }
-
-    // normalize w
-    blas::scal(1.0 / residual, w);
-    // W = A w
-    blas::matvec(A, w, W);
-
     // reset is_singular flag
     if (iter == 0 || is_singular) {
       is_singular = false;
-      Sam.set_row(3);
-      Sam.set_col(3);
-      Sbm.set_row(3);
-      Sbm.set_col(3);
-      wxp.set_row(3);
-      twxp.set_col(3);
-      WXP.set_row(3);
+      Sam.set_row(3 * m);
+      Sam.set_col(3 * m);
+      Sbm.set_row(3 * m);
+      Sbm.set_col(3 * m);
+      wxp.set_row(3 * m);
+      wxp_p.set_row(3 * m);
+      twxp.set_col(3 * m);
+      WXP.set_row(3 * m);
+      WXP_p.set_row(3 * m);
     }
   }
-  blas::copy(x, xinout);
+  for (std::size_t i = 0; i < m; ++i) {
+    view1D<matrix::Dense<T>, T> xinout_i(xinout, i * n, (i + 1) * n);
+    blas::copy(x[i], xinout_i);
+  }
   logger.solver_out();
   return MONOLISH_SOLVER_MAXITER;
 }
 
 template int
 standard_eigen::LOBPCG<matrix::CRS<double>, double>::monolish_LOBPCG(
-    matrix::CRS<double> &A, double &l, vector<double> &x);
+    matrix::CRS<double> &A, vector<double> &l, matrix::Dense<double> &x);
 template int standard_eigen::LOBPCG<matrix::CRS<float>, float>::monolish_LOBPCG(
-    matrix::CRS<float> &A, float &l, vector<float> &x);
+    matrix::CRS<float> &A, vector<float> &l, matrix::Dense<float> &x);
 // template int
 // standard_eigen::LOBPCG<matrix::LinearOperator<double>,
 // double>::monolish_LOBPCG(
@@ -209,7 +317,8 @@ template int standard_eigen::LOBPCG<matrix::CRS<float>, float>::monolish_LOBPCG(
 //     matrix::LinearOperator<float> &A, float &l, vector<float> &x);
 
 template <typename MATRIX, typename T>
-int standard_eigen::LOBPCG<MATRIX, T>::solve(MATRIX &A, T &l, vector<T> &x) {
+int standard_eigen::LOBPCG<MATRIX, T>::solve(MATRIX &A, vector<T> &l,
+                                             matrix::Dense<T> &x) {
   Logger &logger = Logger::get_instance();
   logger.solver_in(monolish_func);
 
@@ -223,9 +332,9 @@ int standard_eigen::LOBPCG<MATRIX, T>::solve(MATRIX &A, T &l, vector<T> &x) {
 }
 
 template int standard_eigen::LOBPCG<matrix::CRS<double>, double>::solve(
-    matrix::CRS<double> &A, double &l, vector<double> &x);
+    matrix::CRS<double> &A, vector<double> &l, matrix::Dense<double> &x);
 template int standard_eigen::LOBPCG<matrix::CRS<float>, float>::solve(
-    matrix::CRS<float> &A, float &l, vector<float> &x);
+    matrix::CRS<float> &A, vector<float> &l, matrix::Dense<float> &x);
 // template int
 // standard_eigen::LOBPCG<matrix::LinearOperator<double>, double>::solve(
 //     matrix::LinearOperator<double> &A, double &l, vector<double> &x);