+++ CHANGELOG.md
@@ -22,12 +22,14 @@ Unreleased
 -----------
 ### Added
 - add 1Dview VML https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/248
+- add benchmarking code in LOBPCG test https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/250
 - add matvec of linearoperator with view1D https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/247
 - add view1D BLAS https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/245
 - add view1D axpy test https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/245
 
 ### Changed
 - generate matrix VML https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/248
+- support multiple eigenpairs in LOBPCG https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/249
 - Use view1D in LOBPCG https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/246
 - support view1D in is_same_size() https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/245